int to double

+++ src/Cash.java
@@ -1,6 +1,6 @@
 public class Cash implements PaymentMethod{
     @Override
-    public void Pay(int amount) {
+    public void Pay(double amount) {
         System.out.println("You have Paid an amount of: "+amount+"$ Successfully");
     }
 }
